fea(api): filter test data from api and refactor
